@@ -24,18 +24,18 @@
   <parent>
     <groupId>org.springframework.boot</groupId>
     <artifactId>spring-boot-starter-parent</artifactId>
-    <version>2.7.4</version>
+    <version>3.2.2</version>
     <relativePath/> <!-- lookup parent from repository -->
   </parent>
 
   <groupId>edu.cnm.deepdive</groupId>
   <artifactId>codebreaker-solitaire-service</artifactId>
-  <version>3.1.1</version>
+  <version>4.1.1</version>
   <name>Codebreaker Solitaire Service</name>
   <description>Code generator &amp; game service</description>
 
   <properties>
-    <java.version>11</java.version>
+    <java.version>21</java.version>
     <maven.build.timestamp.format>yyyy-MM-dd HH:mm z</maven.build.timestamp.format>
   </properties>
 
